backend/requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
openai>=1.35.7
langchain>=0.2.6
langchain-openai>=0.1.7
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
boto3>=1.34.0
faiss-cpu>=1.7.4
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
